Correct CRAM M5 header creation bug with refs out of order.
Note this doesn't happen on the command line, so it's something
specific to the way test/sam.c is written that triggers this.

test/sam.c calls cram_load_reference before the SAM header is known.
That in turn means refs2id hasn't been called and fd->refs array isn't
in sync with the file header (as we have a subset of refs).  Then
cram_write_SAM_hdr md5sums the wrong reference when created SQ M5
tags.

There was already code in the cram_write_SAM_hdr to recall refs2id to
get things in sync, but unfortunately it was after the creation of M5
tags.

What's odd is how this ever worked in test/sam.  Maybe it just didn't
and we never noticed?  It was only spotted here due to the new
validation code.