Make tbx_parse1 work for lines longer than 2Gbytes
Change the type of len, i and b so they can index positions
over 2Gbytes into the input string.  Naturally this is only useful
on 64 bit platforms as the entire input needs to be in memory.

Luckily this function isn't exported, so changing the type of
len does not affect the ABI.