Don't error when making an iterator on a tid not in the index
Instead return one that instantly finishes.

Fixes an edge case (issue #1534) where an index did not include
an entry for a chromosome that was mentioned in the file header
but had no data records.  Normally these would be present but
empty, but it was possible to use the IDX= key in a VCF file
to make an index where the chromosome simply did not appear.  In
this case, rather than an error, we want to return the equivalent
of HTS_IDX_NONE so the iterator produces no data.

Another scenario where this is useful is if you build an index,
and then try to use it immediately without first saving and
reading it back in again.  Such an index will have NULL entries
in bidx[] for any chromosomes with no data.  Again we want to return
an HTS_IDX_NONE iterator if one of those chromosomes is queried.
(This issue didn't usually occur because most programs are loading
in an existing index, and idx_read_core() makes bidx[] entries
for everything even if there's nothing in the index for the
chromosome.)

Note that this changes vcf_loop() in test_view.c so that it
now treats bcf_itr_querys() failures as an error.  The new
behaviour matches sam_loop() and is needed to detect the problem
being fixed here.  All the other tests still work after this change
no nothing was relying on the old behaviour of ignoring the errors.