Add CRAM SQ/M5 header checking when specifying a fasta file. (PR #1522)
* Add CRAM SQ/M5 header checking when specifying a fasta file.

Given the possibility of creating a CRAM which cannot be decoded again
due to specifying a fasta file that differs to the M5 tags specified
in the header, we're now more militant in enforcing these match.

The references are checked on-the-fly as they're used.  For name
sorted data this may have a significant initial cost (the same that we
see when converting from BAM to CRAM without M5 tags in the BAM
header), but for position streaming the cost is only for chromosomes
currently in use.

This means writing a small CRAM file covering a region from just one
chromosome will only incur the CPU cost of validating that one
chromosome, as we're not trying to check the header tags are valid (we
take them at their word) and are instead trying to ensure the fasta
sequence we've been given matches the supplied headers.

* Correct CRAM M5 header creation bug with refs out of order.

Note this doesn't happen on the command line, so it's something
specific to the way test/sam.c is written that triggers this.

test/sam.c calls cram_load_reference before the SAM header is known.
That in turn means refs2id hasn't been called and fd->refs array isn't
in sync with the file header (as we have a subset of refs).  Then
cram_write_SAM_hdr md5sums the wrong reference when created SQ M5
tags.

There was already code in the cram_write_SAM_hdr to recall refs2id to
get things in sync, but unfortunately it was after the creation of M5
tags.

What's odd is how this ever worked in test/sam.  Maybe it just didn't
and we never noticed?  It was only spotted here due to the new
validation code.