Improve API docs for bgzf_mt
Note that this now works on reading as well as writing.

Removed the comment about only when compiled with -DBGZF_MT as this
was misleading implying the user has to adjust CFLAGS.  It's on by
default and there's nothing the user can do to turn it off without
manually editing bgzf.c, in which case they're on their own.

Also document that n_sub_blks now does nothing.

Fixes #1553