Fix a bug with multi-threading and embed_ref=2 on name sorted data
Specifically, this is a multi-threading bug where the
cram_encode_container fails leading to access of freed memory.  This
bug always existed, but we now have a guaranteed way of making the
container encode fail.

Embed_ref=2 computes a consensus and uses that as the reference.  By
design, it cannot work on non-position sorted data, hence a guaranteed
failure for encoding.

(Step 2 will be to do something more sensible than just fail.)