Use non-ref mode when all else fails for CRAM encoding
If we're attempting to use embed_ref=2 because we have alignments in
BAM and no known reference (no M5 tags), and the data is also name
sorted so we can't auto-generate a reference from the consensus and/or
MD:Z tags, then we now just switch to no_ref mode instead of bailing
out.

This isn't ideal, but there is little else that can be done unless the
user modifies their command line or reheaders the input file.

Extra complexities exist as the setting of embed_ref occurs while
encoding a container, which also means it can be adjusted within
threads and synchronously with other containers being encoded.  Hence
we now cache more variables from cram_fd into cram_container.