Add Makefile rule to update the symbol version file
Adds a very basic script to do this.  It expects an htslib.map
file to be already present, assumes that all entries are
in order and that it has a suitable format for the part that
extracts the previous version number.  It extracts any new symbols
from the .so file and appends them to the map, but only if the
HTSlib version number (extracted from PACKAGE_VERSION) has changed.

Apart from for testing, this rule should only be used when building
a release.  Between releases new symbols should be left unversioned
so that binaries linked against a libhts.so will continue to work
if the library is upgraded from one build against develop to the
next release.

After running this rule, the shared library should be rebuilt
so that it picks up the new version names, which makes for a
circular dependency between libhts.so and htslib.map.  This rule
does not attempt to solve that, so the shared library will
currently have to be rebuilt manually.