Add configure (enable|disable)-versioned-symbols options and tests
Default is to try to enable versioned symbols in libhts.so.
This can be overridden using --disable-versioned-symbols.  A
simple test is made to check that the linker option is recognised.
The test uses the htslib.map file in the source - it seems to work
even though the symbols in the map aren't referenced in the test
program being linked.

The Makefile and config.mk.in are adjusted so that the configure
result overrides the Makefile default.