Prevent spurious and random system errors from test_bgzf.c
The test script is expected to fail as it's testing both working and
failing conditions, but the test harness prints up strerror(errno),
which is usually some unrelated event and the text also differs per
platform.

Also, as per 1.16, changed it so the things that are expected to fail
no longer complain.  Given some of the new diagnostics added to
bgzf.c, this requires manipulation of the global htslib log level.