Fix cram_index_query_last function
The cram_index_query_last used by sam_itr_regarray had problems when
dealing with slices whose region is contained within another region.

Fundamentally this is due to the cram_index arrays being contiguous in
memory until a containment is found, at which point the pointers will
be to an entirely different array.  This breaks naive pointer
comparisons.

The cram_index struct already had a "next" field holding the file
offset of the next container.  This has been replaced by e_next
pointing to the next cram_entry struct in file ordering, and
e_next->offset is equivalent to the old "next".  This allows
consumption of the index either as the original nested containment
list or as a traditional linked list.

Also fixed cram_index_query with from != NULL, which similarly was
incorrect before.  We never used this function and it's not public,
but we now use it within the rewrite of cram_index_query_last.

Fixes #1569