Fix crypt4gh redirection
hopen() uses a ':' in `mode` to indicate that there are extra
parameters.  hts_crypt4gh_redirect() needs to add this so that
the "parent" parameter is picked up by the crypt4gh plug-in.

Failing to do this caused it to re-open the file instead of
reusing the existing file handle - which worked for regular files
but not for things like pipes or htsget.