Fix bug where bin number would overflow in hts_itr_query
When searching for `max_off`, hts_itr_query() looks for a bin
to the right of the end of the region.  For whole chromosomes,
this would be HTS_POS_MAX, which is far beyond the maximum
bin position supported.  The `bin` calculation overflowed leading
to much implementation-defined behaviour as it explored various
negative bin numbers.

Fix by limiting `end` to the maximum value supported by the index.