Make reg2bins faster on whole-chromosome queries
It's faster to iterate through the index hash table when
requesting whole chromosomes, compared to the old behaviour
of iterating through all the bins that could exist and looking
them up in the hash table to see if they're present.  The latter
method works better for narrow ranges though, so we choose which
to use based on the number of bins covering the range compared
to the number in the index.