Fix fd_seek on pipes on modern MinGW releases.
MinGW 12.x started returning non-zero values from lseek when the fd
is a pipe.  This is unhelpful and it breaks bgzf_check_EOF as seeking
to the end is actually seeking to the end of the pipe memory buffer,
causing invalid EOFs.  (This breaks bcftools CI tests.)

Fixes samtools/bcftools#1901

Co-authored-by: John Marshall <jmarshall@hey.com>