Protect against overly large containers.
It's possible to construct CRAM containers that are extremely large
such that building a block of BAM records representing the container
overflows due to the size of the combined aux fields.

We could change how we construct blocks of data, and work at a more
individual read level, but realistically it's just not good form to be
handling arbitrarily large containers as they may cause excessive
memory issues which brings its own denial attacks.

POTENTIAL SECURITY ISSUE: Note the previous could overflow cr->aux,
which then went negative and caused negative offsets to be passed to
memcpy.  This would lead to a crash.  I cannot see a way to get this
to not crash and hence leak data, but it could form a denial of
service on a remote server using htslib.

.