Adjust comments in probaln_glocal()
Adds a comment explaining that the f[] and b[] arrays count
positions from 1, allowing 0 to be used to more easily handle
the edges of the alignment matrix.

Changes the comment explaining the line:
           if (u < 3 || u >= i_dim) continue;
used in some of the loops over f[] and b[].  While it does
prevent overstepping the array boundaries, its main
function is to select the parts over which the scores have
previously been calculated.  A change in 5d7a7823 to fix
excess memory usage got the high end slightly wrong (using
i_dim - 3).  When the query sequence length was less than the
band width, this could lead to the last column being
incorrectly missed out from parts of the calculation.