Add synced reader region tests, and move no-index tests
Add some tests to exercise the --regions / --targets synced reader
options.  Currently this only includes tests for the chromosomes
with [:-] characters in the name, but it could be expanded easily
to do others.  Test files have been borrowed from pull request
samtools/bcftools#1938.

Move the synced reader no-index tests from test-bcf-sr.pl to
test.pl.  The former isn't a good place for them as it gets
called 10 times, but the no-index test only needs to run once.
It also allows the code running the test to be simplified a bit.

Also fix the exit code on test-bcf-sr.pl failure from -1 to 1.

Co-authored-by: Petr Danecek <pd3@sanger.ac.uk>