The first stage of vcf_parse_format speed improvements.
This simply turns the monolithic vcf_parse_format functions into a
series of numbers sub-functions whose primary is to localise the
variables to that code block and to make it easier to see the
structure of the tasks being performed.

There is no code optimisation here and the main algorithm is
unchanged, so this is just moving of code from 1 function to multiple
functions.  However it makes the next commit easier to understand as
we're not trying to see a delta mixed in with restructuring.

An unexpected consequence however of making the variables local to
their blocks is that it also speeds up the code.

The first step in separating this code into functions was simply
adding curly braces around each code segment and moving the
function-global variables into their respective blocks.  The
before/after benchmarkjs on 100,000 lines of a multi-sample G1K VCF
are ("perf stat" cycle counts):

        ORIG            LOCALISED
gcc7    29335942870     27353443704
gcc13   31974757094     31452452908
clang13 31290989382     29508665020

Benchmarked again after moving to actual functions, but the difference
was tiny in comparison.)