Prevent dangling hrec pointer after bcf_hdr_add_hrec() failure
Borrow a bit of code from bcf_hdr_remove() that removes hrec
pointers from the hdr->dict[] dictionaries, turn it into
bcf_hdr_unregister_hrec() and use it to clean up the dictionary
should bcf_hdr_add_hrec() fail.

There's actually only one place in bcf_hdr_add_hrec() where this
needs to be called.  In all other paths returning an error, either
the hrec type is not one that needs to be cleaned up, or hrec
will not have been added to the dictionary.

bcf_hdr_remove() is updated to call the new function when it's
removing all lines of a given type.  The code handling lines
with a specific key is unchanged as is that case it already has
the key to look up in the dictionary and so doesn't need to
hunt for it in the header record.