Cache key header lengths.
This is a major improvement as we're not continually calling strlen on
the keys every time we go from a key ID to a string.

We have to do it at the last point (writing) as some tools
(e.g. bcftools annotate) modify the header without setting it to dirty
and without calling bcf_hdr_sync.

Also add a specialised version of bcf_fmt_array for array length of 1,
which is a very common scenario, and change most kputc calls to kputc_
as we don't need to keep nul-terminating while constructing the
string.

Optimised bcf_unpack_info_core1 a bit too.