Revert most of the vcf_parse_info improvements.
The two we keep are the internal while loop to find the next ; or =
instead of iterating back in the outer for loop, and memcmp instead of
strcmp for "END".

The strchr changes do help glibc on excessively long INFO tokens, seen
in the GIAB truth set and GNOMAD files, but they have no impact on
most mainstream VCF outputs.  Furthermore, other C libraries, such as
MUSL, are considerably slowed down by the use of strchr.  Hence this
isn't a particularly robust or warranted change.