Speed up kputd by approx 130%.
We were on a reasonable track before, but given this is emulating an
sprintf %g format which has 6 significant digits, converting the full
"i" integer to ASCII and then only printing the first 6 is redundant.

Instead of multiplying d by a fixed amount and adding varying amounts
for rounding, we multiply by a varying amount and add a fixed .5 for
rounding, in doing so also forcing the integer to now always be 6
digits (bar some very rare rounding issues of i == 1000000).

I've tested this with some 23 million numbers from 0.0001 to 999999
and found no differences.

The use of __builtin_clz can speed this up.  We'd need to check
"if (__builtin_clz(1+(int)d) < 31)" and jump past the fractions to the
"d < 10" (&& >= 1) case, but in tests it looks to only alter speed by
a couple of percent, so it's probably not worth the extra complexity
of compiler conditionals.