Fix crypt4gh redirection
Commit b8cf3463 was a fix to make crypt4gh redirection work
with htsget.  Unfortunately it had a side effect of breaking
crypt4gh redirection on other types of input because
hts_hopen() makes the crypt4gh hFILE use the input hFILE as a
parent, but then goes on to close it because it had been
redirected.  (Before b8cf3463 it didn't matter because crypt4gh
was incorrectly reopening the file instead of using the handle
passed in to it).

Fixing this is a bit tricky because crypt4gh and htsget
redirections work in rather different ways.  For htsget
we get a completely new handle, and want to close the original
because it's no longer needed.  For crypt4gh the original handle
is wrapped by the new one and needs to stay open.  There are
added complications should an error occur after a redirection,
as then hts_hopen() returns NULL and expects the caller to close
the hFILE that was passed to it.

The solution is to add a preserve bit to the hFILE struct
bitfield, which makes hclose() and hclose_abruptly() ignore the
file (although hclose() will still flush it).  This bit is then
set on the input handle by hts_hopen() if it's wrapped inside
a cyrpt4gh handle, allowing error handling to clean up any
handles that happen to have been made by hts_hopen() while leaving
the original one intact.  File cleanup for htsget redirects is
also changed slightly so that the correct handles get closed on a
successful return.