Fix virtual offset adjustment when indexing on-the-fly
When indexing on-the-fly, the virtual offset of the end of
each read gets passed to hts_idx_push().  As HTSlib starts
a new BGZF block if the next read doesn't fit, the stored
offset needs to be adjusted to point to the start of the
next block instead of the end of the old one.  (While the
two locations effectively map to the same uncompressed
offset, pointing to the end of a block causes problems for
HTSlib before 1.11, and HTSJDK.)

When multi-threading, this adjustment was done by calling
bgzf_idx_amend_last() from the main thread.  However, if the
data was written out too promptly, the function could be called
too late preventing the adjustment from being applied.  To fix
this, the adjustment is now moved to bgzf_idx_flush() which is
called by the writer thread, ensuring that it is always applied
on time.

The internal function bgzf_idx_amend_last() is removed as it's no
longer needed.

Fixes samtools/samtools#1861 (samtools merge --write-index
  results in Invalid file pointer [from HTSJDK] in rare cases
  whereas samtools index fixes the problem)