Fix hfile_libcurl small seek bug
Ensure that `to_skip` is updated correctly when handling a
small seek by skipping data in libcurl_read().  In particular,
if the number of bytes returned by wait_perform() exactly
matches the number left to skip, `to_skip` needs to be set to zero,
and wait_perform() needs to be called again to get some data for
the function to return.

Co-Authored-By: Rob Davies <rmd+git@sanger.ac.uk>