Add test case compiling the public headers as C++
Add a C++ test file that includes all the public HTSlib headers,
to verify that the headers compile cleanly when compiled as C++.

Rather than adding configury to find a C++ compiler, take advantage
of Clang and GCC's drivers using C++ when given a .cpp file. As this
may not be true in general, add this test to maintainer-check only.