Fix a couple of unused value warnings when built with NDEBUG
The warnings appeared due to the values only being used for
assert() statements.  The one in cram_stats_encoding() is
just checking some counts, so the variable is just annotated
with HTS_UNUSED to silence the warning.

The assert() in cram_index_build() could actually be triggered
with a suitably-adjusted cram file.  As it therefore detects
bad inputs, it is replaced with code to write a better error
message and return the documented value for read errors in this
function.