Avoid a NULL pointer deref when erroring writing CRAM to stdout
If we write to stdout and get an error, samtools error printing does
an hts_flush of stdout, followed by an hts_close which then also does
its own flush.

We weren't resetting fd->ctr to NULL, causing a second attempt to
encode the container.  However we've already discarded the c->bams
array.  We double check that too for safety, but it should be
unnecessary now we correctly free and nullify the container in
cram_flush rather than relying on it in cram_close, so multiple
flushes don't trigger this bug.