Fix memory leaks on failed CRAM encode.
The c->bams list is normally moved over to spare_bams list on success,
but on encode failure it's still there so we need to free this.

Additionally failure to encode a set of aux tags leaves some aux
blocks in use in the container still which haven't yet been migrated
over to s->aux_blocks.