Permit embed_ref=2 mode to be reenabled after using no_ref.
Commit 0ab97e3 limited memory use for excessively large spanning
slices.  This could be extremely sparse data, in which case embedding
a reference isn't helpful, or it could be deep data which a huge gap
(eg spanning a centromere).  Our maximum gap is still big - approx
100Mb span - so this is unlikely to happen.  However if we do trigger
it, we never get the option of going back from no-ref to embed-ref=2
(ie own consensus) again.

This commit adds a counter on the number of times we've failed to
generate a reference in recent times, incrementing or decrementing
accordingly.  Once it hits a certain level we give up, as we did
before.  So for example name sorting data will try 5 times and then
never attempt to use embed_ref again.