Fix buffer read overrun in cram_encode_aux.
This is triggered when the aux data hasn't been validated and contains
a partial buffer.  Eg a non-null terminated string, or an 'i' integer
with < 4 bytes remaining in the buffer.

This is a bit involved as the htslib API also lacks bounds checking in
the utility functions like bam_aux2i.  I added a local bounds checking
equivalent, but in time maybe this znd other similar functions should
be moved to become an official part of the API.