Ensure no-stored-sequence reads are counted in container size
In CRAM, sequence bases have to be stored for certain read
features, even if the input record has SEQ set to "*" (i.e. no
sequence stored).  In such cases, the sequence becomes a string
of "N"s, which need to be accounted for when calculating how many
items are put into a slice.  Failing to do this could lead to
some data structures overflowing with certain unusual inputs.
The number of Ns is set by the query-consuming operators in
the CIGAR string, which can become quite large if you have
lots of long insertions.

A limit is also set on the apparent sequence length of these
records to prevent problems with excessive time and memory use
when trying to encode them.  This also prevents issues with
integer overflow if the apparent sequence length exceeded INT_MAX.

In cram_encode_container() an extra check is added to ensure
at least one record is present, to fix a crash that could
occur if cram_put_bam_seq() bailed out with an error.

Credit to OSS-Fuzz
Fixes oss-fuzz 64557