Fix over-lenient check in bam_aux_first() / bam_aux_next()
After calling these, bam_aux_get() expects three bytes to be
available (two before the returned pointer).  While a test
to prevent running off the end of the aux data was present, it
didn't account for incrementing the returned pointer by 2.  The
new version ensures that the expected data is available.

The bug could lead to a read of up to two bytes beyond the end
of the aux data in bam_aux_get().  As a later call to skip_aux()
would have correctly detect that the tag is invalid, bam_aux_get()
should still have returned NULL, making it difficult to exploit
the out-of-bounds access.

Credit to OSS-Fuzz
Fixes oss-fuzz 65820